requirements.txt
Django==2.1
django-cors-headers==3.2.0
djangorestframework==3.10.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
Jinja2==2.10.3
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
openpyxl==3.0.2
parameterized==0.7.0
parsel==1.5.2
Pillow==6.2.1
pipenv==2018.11.26
Protego==0.1.15
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
pygal==2.4.0
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
virtualenv==16.7.8
virtualenv-clone==0.5.3
w3lib==1.21.0
zope.interface==4.7.1
